Tpetra: Harness: Add for_each header; make generic
@trilinos/tpetra

1. Split off for_each into its own header.

2. Make for_each generic, with an implementation class that we can
   specialize for other types besides Tpetra::MultiVector and
   Tpetra::Vector.

3. Separate the MultiVector and Vector specializations of the
   implementation of for_each into a separate header.
