+++ packages/tpetra/core/test/MultiVector/Harness.cpp
@@ -48,434 +48,12 @@
 #include "Tpetra_Vector.hpp"
 #include "Tpetra_withLocalAccess.hpp"
 #include "Tpetra_withLocalAccess_MultiVector.hpp"
+#include "Tpetra_for_each.hpp"
+#include "Tpetra_for_each_MultiVector.hpp"
 #include "Tpetra_Details_Behavior.hpp"
 #include "Kokkos_ArithTraits.hpp"
 #include "Teuchos_DefaultSerialComm.hpp"
 #include "Teuchos_TypeNameTraits.hpp"
-
-namespace Tpetra {
-
-  namespace Details {
-    ////////////////////////////////////////////////////////////
-    // Implementation details of for_each
-    ////////////////////////////////////////////////////////////
-
-    // Implementation detail of for_each (see below).  Given a Kokkos
-    // execution space on which the user wants to run the for_each,
-    // and a memory space in which the MultiVector's data live,
-    // determine the memory space that for_each should use in its
-    // withLocalAccess call.
-    template<class ExecutionSpace, class MemorySpace>
-    using for_each_memory_space =
-      typename std::conditional<
-        Kokkos::SpaceAccessibility<
-          ExecutionSpace,
-          typename MemorySpace::memory_space>::accessible,
-        typename MemorySpace::memory_space,
-        typename ExecutionSpace::memory_space>::type;
-
-    // for_each uses this for the loop body of a parallel_for or
-    // parallel_reduce over the rows of a Tpetra::MultiVector with
-    // constant stride and multiple columns.
-    template<class ViewType,
-             class InnerLoopBodyType,
-             class LocalIndexType>
-    struct MultiVectorOuterForEachLoopBody {
-      static_assert (static_cast<int> (ViewType::Rank) == 2,
-                     "ViewType must be a rank-2 Kokkos::View.");
-      MultiVectorOuterForEachLoopBody (const ViewType& X_lcl,
-                                       InnerLoopBodyType f) :
-        X_lcl_ (X_lcl), f_ (f)
-      {}
-      KOKKOS_INLINE_FUNCTION void
-      operator () (const LocalIndexType i) const
-      {
-        const LocalIndexType numCols =
-          static_cast<LocalIndexType> (X_lcl_.extent (1));
-        for (LocalIndexType j = 0; j < numCols; ++j) {
-          f_ (X_lcl_(i,j), i, j);
-        }
-      };
-      ViewType X_lcl_;
-      InnerLoopBodyType f_;
-    };
-
-    // for_each uses this for the loop body of a parallel_for or
-    // parallel_reduce over the rows of a Tpetra::Vector (or each
-    // column of a Tpetra::MultiVector, if the Tpetra::MultiVector has
-    // nonconstant stride or only a single column).
-    template<class ViewType,
-             class InnerLoopBodyType,
-             class LocalIndexType>
-    struct VectorOuterForEachLoopBody {
-      static_assert (static_cast<int> (ViewType::Rank) == 1,
-                     "ViewType must be a rank-1 Kokkos::View.");
-      VectorOuterForEachLoopBody (const ViewType& X_lcl,
-                                  InnerLoopBodyType f) :
-        X_lcl_ (X_lcl), f_ (f)
-      {}
-      KOKKOS_INLINE_FUNCTION void
-      operator () (const LocalIndexType i) const
-      {
-        f_ (X_lcl_(i), i, LocalIndexType (0));
-      };
-      ViewType X_lcl_;
-      InnerLoopBodyType f_;
-    };
-
-    // Distinguish between functions that take (scalar&, index,
-    // index), (scalar&, index), and (scalar&).
-    enum class EMultiVectorForEachFuncArgs {
-      SCALAR,
-      SCALAR_ROWINDEX,
-      SCALAR_ROWINDEX_COLINDEX,
-      ERROR
-    };
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    constexpr bool
-    isScalarIndexIndexFunction ()
-    {
-      using UFT = UserFunctionType;
-      using IST = ImplScalarType;
-      using LIT = LocalIndexType;
-      using func_type_1 =
-        std::function<void (IST&, const LIT, const LIT)>;
-      using func_type_2 = std::function<void (IST&, LIT, LIT)>;
-
-      return std::is_convertible<UFT, func_type_1>::value ||
-        std::is_convertible<UFT, func_type_2>::value;
-    }
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    constexpr bool
-    isScalarIndexFunction ()
-    {
-      using UFT = UserFunctionType;
-      using IST = ImplScalarType;
-      using LIT = LocalIndexType;
-      using func_type_1 = std::function<void (IST&, const LIT)>;
-      using func_type_2 = std::function<void (IST&, LIT)>;
-
-      return std::is_convertible<UFT, func_type_1>::value ||
-        std::is_convertible<UFT, func_type_2>::value;
-    }
-
-    template<class UserFunctionType, class ImplScalarType>
-    constexpr bool
-    isScalarFunction ()
-    {
-      using func_type_1 = std::function<void (ImplScalarType&)>;
-
-      return std::is_convertible<UserFunctionType, func_type_1>::value;
-    }
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    constexpr EMultiVectorForEachFuncArgs
-    getMultiVectorForEachFuncArgs ()
-    {
-      using UFT = UserFunctionType;
-      using IST = ImplScalarType;
-      using LIT = LocalIndexType;
-
-      return isScalarIndexIndexFunction<UFT, IST, LIT> () ?
-        EMultiVectorForEachFuncArgs::SCALAR_ROWINDEX_COLINDEX :
-        (isScalarIndexFunction<UFT, IST, LIT> () ?
-         EMultiVectorForEachFuncArgs::SCALAR_ROWINDEX :
-         (isScalarFunction<UFT, IST> () ?
-          EMultiVectorForEachFuncArgs::SCALAR :
-          EMultiVectorForEachFuncArgs::ERROR));
-    }
-
-    // Functor that MultiVectorOuterForEachLoopBody or
-    // VectorOuterForEachLoopBody uses.  This functor in turn wraps
-    // the user's function given to for_each.  We have different cases
-    // for whether the user's function takes (scalar&, local row
-    // index, local column index), (scalar&, local row index), or
-    // (scalar&).
-    //
-    // The point of MultiVectorInnerForEachLoopBody is so that
-    // Tpetra::for_each only needs one implementation, but can work
-    // with three different kinds of user functions: (scalar&, local
-    // row index, local column index), (scalar&, local row index), and
-    // (scalar&).
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType,
-             const EMultiVectorForEachFuncArgs argsType =
-               getMultiVectorForEachFuncArgs<UserFunctionType,
-                                             ImplScalarType,
-                                             LocalIndexType> ()>
-    struct MultiVectorInnerForEachLoopBody {
-      static_assert (argsType != EMultiVectorForEachFuncArgs::ERROR,
-                     "Please report this bug to the Tpetra developers.");
-    };
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    struct MultiVectorInnerForEachLoopBody<
-      UserFunctionType, ImplScalarType, LocalIndexType,
-      EMultiVectorForEachFuncArgs::SCALAR_ROWINDEX_COLINDEX>
-    {
-      MultiVectorInnerForEachLoopBody (UserFunctionType f) :
-        f_ (f) {}
-      KOKKOS_INLINE_FUNCTION void
-      operator () (ImplScalarType& x_ij,
-                   const LocalIndexType i,
-                   const LocalIndexType j) const
-      {
-        f_ (x_ij, i, j);
-      };
-      UserFunctionType f_;
-    };
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    struct MultiVectorInnerForEachLoopBody<
-      UserFunctionType, ImplScalarType, LocalIndexType,
-      EMultiVectorForEachFuncArgs::SCALAR_ROWINDEX>
-    {
-      MultiVectorInnerForEachLoopBody (UserFunctionType f) :
-        f_ (f) {}
-      KOKKOS_INLINE_FUNCTION void
-      operator () (ImplScalarType& x_ij,
-                   const LocalIndexType i,
-                   const LocalIndexType /* j */) const
-      {
-        f_ (x_ij, i);
-      };
-      UserFunctionType f_;
-    };
-
-    template<class UserFunctionType,
-             class ImplScalarType,
-             class LocalIndexType>
-    struct MultiVectorInnerForEachLoopBody<
-      UserFunctionType, ImplScalarType, LocalIndexType,
-      EMultiVectorForEachFuncArgs::SCALAR>
-    {
-      MultiVectorInnerForEachLoopBody (UserFunctionType f) :
-        f_ (f) {}
-      KOKKOS_INLINE_FUNCTION void
-      operator () (ImplScalarType& x_ij,
-                   const LocalIndexType /* i */,
-                   const LocalIndexType /* j */) const
-      {
-        f_ (x_ij);
-      };
-      UserFunctionType f_;
-    };
-
-    // The implementation of for_each uses the result of
-    // makeMultiVectorForEachLoopBody or makeVectorForEachLoopBody as
-    // the functor in a parallel_for over the local rows of the
-    // Tpetra::(Multi)Vector.
-
-    template<class ViewType,
-             class UserFunctionType,
-             class LocalIndexType>
-    MultiVectorOuterForEachLoopBody<
-      ViewType,
-      MultiVectorInnerForEachLoopBody<
-        UserFunctionType,
-        typename ViewType::non_const_value_type,
-        LocalIndexType>,
-      LocalIndexType>
-    makeMultiVectorForEachLoopBody (const ViewType& X_lcl,
-                                    UserFunctionType f,
-                                    const LocalIndexType /* numCols */)
-    {
-      using return_type = typename ViewType::non_const_value_type;
-      using inner_loop_body_type =
-        MultiVectorInnerForEachLoopBody<
-          UserFunctionType, return_type, LocalIndexType>;
-      using outer_loop_body_type =
-        MultiVectorOuterForEachLoopBody<
-          ViewType, inner_loop_body_type, LocalIndexType>;
-      return outer_loop_body_type (X_lcl, inner_loop_body_type (f));
-    }
-
-    template<class ViewType,
-             class UserFunctionType,
-             class LocalIndexType>
-    VectorOuterForEachLoopBody<
-      ViewType,
-      MultiVectorInnerForEachLoopBody<
-        UserFunctionType,
-        typename ViewType::non_const_value_type,
-        LocalIndexType>,
-      LocalIndexType>
-    makeVectorForEachLoopBody (const ViewType& X_lcl,
-                               UserFunctionType f,
-                               const LocalIndexType /* numCols */)
-    {
-      using return_type = typename ViewType::non_const_value_type;
-      using inner_loop_body_type =
-        MultiVectorInnerForEachLoopBody<
-          UserFunctionType, return_type, LocalIndexType>;
-      using outer_loop_body_type =
-        VectorOuterForEachLoopBody<
-          ViewType, inner_loop_body_type, LocalIndexType>;
-      return outer_loop_body_type (X_lcl, inner_loop_body_type (f));
-    }
-
-    // Implementation of for_each (see below).
-    template<class ExecutionSpace,
-             class TpetraMultiVectorType,
-             class UserFunctionType>
-    struct ForEach {
-      static void
-      for_each (const char kernelLabel[],
-                ExecutionSpace execSpace,
-                TpetraMultiVectorType& X,
-                UserFunctionType f)
-      {
-        using Teuchos::TypeNameTraits;
-        using std::endl;
-        using MV = TpetraMultiVectorType;
-        using preferred_memory_space =
-          typename MV::device_type::memory_space;
-        using memory_space = Details::for_each_memory_space<
-          ExecutionSpace, preferred_memory_space>;
-        using LO = typename MV::local_ordinal_type;
-        using range_type = Kokkos::RangePolicy<ExecutionSpace, LO>;
-
-        const int myRank = X.getMap ()->getComm ()->getRank ();
-        const bool verbose = ::Tpetra::Details::Behavior::verbose ();
-        if (verbose) {
-          std::ostringstream os;
-          os << "Proc " << myRank << ": Tpetra::for_each:" << endl
-             << " kernelLabel: " << kernelLabel << endl
-             << " ExecutionSpace: "
-             << TypeNameTraits<ExecutionSpace>::name () << endl
-             << " memory_space: "
-             << TypeNameTraits<memory_space>::name () << endl;
-          std::cerr << os.str ();
-        }
-
-        memory_space memSpace;
-        if (X.getNumVectors () == size_t (1) || ! X.isConstantStride ()) {
-          const size_t numVecs = X.getNumVectors ();
-          for (size_t j = 0; j < numVecs; ++j) {
-            auto X_j = X.getVectorNonConst (j);
-            // Generic lambdas need C++14, so we must use arg_type here.
-            using arg_type =
-              with_local_access_function_argument_type<
-                decltype (readWrite (*X_j).on (memSpace))>;
-            withLocalAccess
-              ([=] (const arg_type& X_j_lcl) {
-                auto loopBody =
-                  Details::makeVectorForEachLoopBody (X_j_lcl, f, LO (1));
-                Kokkos::parallel_for
-                  (kernelLabel,
-                   range_type (execSpace, 0, X_j_lcl.extent (0)),
-                   loopBody);
-              }, readWrite (*X_j).on (memSpace));
-          }
-        }
-        else {
-          // Generic lambdas need C++14, so we must use arg_type here.
-          using arg_type =
-            with_local_access_function_argument_type<
-              decltype (readWrite (X).on (memSpace))>;
-          withLocalAccess
-            ([=] (const arg_type& X_lcl) {
-              const LO numCols = static_cast<LO> (X_lcl.extent (1));
-              auto loopBody =
-                Details::makeMultiVectorForEachLoopBody (X_lcl, f, numCols);
-              Kokkos::parallel_for
-                (kernelLabel,
-                 range_type (execSpace, 0, X_lcl.extent (0)),
-                 loopBody);
-            }, readWrite (X).on (memSpace));
-        }
-      }
-    };
-  } // namespace Details
-
-  /// \brief Apply a function entrywise to each entry of a
-  ///   Tpetra::MultiVector or Tpetra::Vector, analogously to
-  ///   std::for_each.
-  ///
-  /// The function f may have any of the following forms:
-  /// <ul>
-  /// <li> Takes the current entry in read-write fashion as
-  ///   <tt>impl_scalar_type&</tt>, the local row index as LO, and the
-  ///   local column index as LO; </li>
-  ///
-  /// <li> Takes the current entry in read-write fashion as
-  ///   <tt>impl_scalar_type</tt> and the local row index as LO; or
-  ///   </li>
-  ///
-  /// <li> Takes the current entry in read-write fashion as
-  ///   <tt>impl_scalar_type</tt>. </li>
-  /// </ul>
-  ///
-  /// <tt>impl_scalar_type</tt> is a public typedef in
-  /// Tpetra::MultiVector and Tpetra::Vector.  When
-  /// <tt>scalar_type</tt> is <tt>std::complex<T></tt>, then
-  /// <tt>impl_scalar_type</tt> is <tt>Kokkos::complex<T></tt>.
-  /// Otherwise, <tt>scalar_type</tt> and <tt>impl_scalar_type</tt>
-  /// are the same.
-  ///
-  /// \param kernellabel [in] Kernel label for Kokkos Profiling.
-  /// \param execSpace [in] Kokkos execution space on which to run.
-  /// \param X [in/out] MultiVector or Vector to modify.
-  /// \param f [in] Function to apply to each entry of X.
-  ///
-  /// There is no overload without a kernel label.  You WILL label
-  /// your kernels.  If you bother to use Kokkos, that means you care
-  /// about performance, so we need to be able to measure it.
-  template<class SC, class LO, class GO, class NT,
-           class UserFunctionType,
-           class ExecutionSpace>
-  void
-  for_each (const char kernelLabel[],
-            ExecutionSpace execSpace,
-            Tpetra::MultiVector<SC, LO, GO, NT>& X,
-            UserFunctionType f)
-  {
-    using MV = Tpetra::MultiVector<SC, LO, GO, NT>;
-    using impl_type =
-      Details::ForEach<ExecutionSpace, MV, UserFunctionType>;
-    impl_type::for_each (kernelLabel, execSpace, X, f);
-  }
-
-  /// \brief Overload of for_each (see above) that runs on X's
-  ///   default Kokkos execution space.
-  ///
-  /// \param kernellabel [in] Kernel label for Kokkos Profiling.
-  /// \param X [in/out] MultiVector to modify.
-  /// \param f [in] Function to apply entrywise to X (could have
-  ///   different signatures; see above).
-  ///
-  /// There is no overload without a kernel label.  You WILL label
-  /// your kernels.  If you bother to use Kokkos, that means you care
-  /// about performance, so we need to be able to measure it.
-  template<class SC, class LO, class GO, class NT,
-           class UserFunctionType>
-  void
-  for_each (const char kernelLabel[],
-            Tpetra::MultiVector<SC, LO, GO, NT>& X,
-            UserFunctionType f)
-  {
-    using MV = Tpetra::MultiVector<SC, LO, GO, NT>;
-    using execution_space = typename MV::device_type::execution_space;
-    using impl_type =
-      Details::ForEach<execution_space, MV, UserFunctionType>;
-    execution_space execSpace;
-    impl_type::for_each (kernelLabel, execSpace, X, f);
-  }
-
-} // namespace Tpetra
 
 namespace { // (anonymous)
